feat: add isActive field to the Course model
The isActive field will now default to false,
and as soon as a user deploys a smart contract for this course,
this field will change to true and will appear in the course store.

Also create a new endpoint for getting all created courses by userId.
Add userId to the sending web app url for showing all created courses by user