feat: remove unnecessary libraries
remove cloudinary, redis, multer as images are now sent with a ready-made link
and there is no need to upload them to cloud storage.
Change userId type on BigInt;
remove imagePublicId from models;
remove model Video;
add videoUrl to the model Lesson.
Remove unnecessary code for listeners bot when loading an image because it is not needed