feat: add ability to save image for lesson in Cloudinary
lesson creation:
If the image comes in by url, it's just saved in the database.
If the image comes as a file, it is saved in Cloudinary and its url and publicId are saved in the database.

lesson editing:
if an image comes in by url, the imageUrl is replaced with the new url and the imagePublicId is deleted
if a file arrives, the old image is deleted from Cloudinary and the new image is saved there and its imageUrl and imagePublicId are saved in the database.
if the isRemoveImage field arrives, the image is cleared from Cloudinary and all fields for the course image are also cleared in the database.