feat: add new field hasAcceptedTerms
now the new hasAcceptedTerms field is received in the monitorContract endpoint.
If it is not received or false, the contract is not checked and no points are credited.
This field is then stored in the Course and CoursePurchase table.
This field indicates whether the user agrees to the terms of the contract:
- For Course, the contract is between the seller and the marketplace;
- For Purchase - the contract between the customer and the seller