feat: create a new endpoint to get user data
also change update function in the user service to check
if the email has changed from the one that was saved from the one that came to this service.
Added a new isCodeResend field to the update function.
These 2 changes were made in order to understand whether the new code should be resent to email
or not and how to save the data correctly.

Also hasAcceptedTerms is now an optional field
and even if it is not sent to the backend,
the monitorContract function will still continue to work