feat: send video through bot
It is now possible to send videos through the bot instead of the frontend.
This improves the user experience as there is no need to wait for the video to fully load.
There is a problem: we can only get the file_path here: await bot.getFile(fileId);
But this won't work if the file is bigger than 20 or 50MB.
So this commit is left for further development and will not be implemented in the main branch for now